Individual

+++ src/main/java/com/example/books4trade/controllers/BookReviewController.java
@@ -2,11 +2,13 @@ package com.example.books4trade.controllers;
 
 
 import com.example.books4trade.models.BookReview;
+import com.example.books4trade.models.OwnedBook;
 import com.example.books4trade.models.User;
 import com.example.books4trade.repositories.BookRepository;
 import com.example.books4trade.repositories.BookReviewRepository;
 import com.example.books4trade.repositories.UserRepository;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -19,13 +21,13 @@ import java.util.List;
 public class BookReviewController {
 
     private final BookReviewRepository bookReviewDao;
-    private final BookRepository bookDao;
-    private final UserRepository userDao;
+    private final BookRepository booksDao;
+    private final UserRepository usersDao;
 
     public BookReviewController(BookReviewRepository bookReviewDao, BookRepository bookDao, UserRepository userDao) {
         this.bookReviewDao = bookReviewDao;
-        this.bookDao = bookDao;
-        this.userDao = userDao;
+        this.booksDao = bookDao;
+        this.usersDao = userDao;
     }
 
     @GetMapping("/reviews")
@@ -40,6 +42,9 @@ public class BookReviewController {
         return "reviews/index";
     }
 
+
+
+//
     @GetMapping("/reviews/{id}")
     public String showIndividualReview (@PathVariable long id, Model model) {
         BookReview individualReview = bookReviewDao.getById(id);
@@ -48,18 +53,43 @@ public class BookReviewController {
         return "reviews/individual-review";
     }
 
-    @GetMapping("/reviews/create")
-    public String showReviewForm(Model model) {
-        model.addAttribute("bookReview", new BookReview());
+
+
+
+    @GetMapping("/reviews/{id}/create")
+    public String showReviewForm(Model model, @PathVariable long id) {
+        model.addAttribute("book", booksDao.getById(id));
+
         return "/reviews/create";
     }
 
-    @PostMapping("/reviews/create")
-    public String createReview (@ModelAttribute BookReview bookReview) {
-//        bookReview.setUser((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
-         bookReviewDao.save(bookReview);
-         return "redirect:/reviews";
+//
+
+//    edited by mike
+    @PostMapping("/reviews/{id}/create")
+    public String createReview (@RequestParam(name= "rating") long rating, @RequestParam(name="body") String body, @RequestParam (name= "title") String title, @PathVariable long id) {
+        BookReview bookReview = new BookReview();
+        bookReview.setBook(booksDao.getById(id));
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        bookReview.setBody(body);
+        bookReview.setTitle(title);
+        bookReview.setRating(rating);
+        bookReview.setUser(currentUser);
+        //        bookReview.setUser((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
+
+         bookReview = bookReviewDao.save(bookReview);
+         return "redirect:/reviews/" +bookReview.getId();
     }
+
+//    this one change url
+//    @GetMapping("/reviews/{id}")
+//    public String showReviewOfBook(@PathVariable long id, Model model){
+//        model.addAttribute("review", bookReviewDao.getById(id));
+//        model.addAttribute("allReviews", bookReviewDao.findBookReviewsBy(booksDao.getById(id)));
+//        return "/reviews/individual-review";
+//    }
+//
+
 
     @GetMapping("reviews/{id}/edit")
     public String editForm(@PathVariable long id, Model model) {
@@ -67,7 +97,23 @@ public class BookReviewController {
         return "/reviews/edit";
     }
 
-//    @PostMapping("/reviews/{id}/edit"){
+
+    @PostMapping("/reviews/{id}/edit")
+        public String submitEdit(@PathVariable long id, @RequestParam (name="rating") String rating, @RequestParam (name="body") String body, @RequestParam(name="title") String title,@ModelAttribute BookReview editedReview){
+        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+            BookReview reviewToEdit = bookReviewDao.getById(editedReview.getId());
+            if(reviewToEdit.getUser().getId() == currentUser.getId()){
+                reviewToEdit.setUser(currentUser);
+                reviewToEdit.setTitle(title);
+                reviewToEdit.setBody(body);
+                reviewToEdit.setRating(rating);
+                reviewToEdit = bookReviewDao.save(reviewToEdit);
+            }
+
+            return "redirect:/reviews" + id + "/user/" + reviewToEdit.getId();
+        }
+
+//        @PostMapping("/reviews/{id}/edit"){
 //        public String submitEdit(@ModelAttribute BookReview editedReview, @PathVariable long id) {
 //            BookReview bookToEdit = bookReviewDao.getById(id);
 //            bookToEdit.setTitle(bookToEdit.getTitle());
@@ -77,23 +123,45 @@ public class BookReviewController {
 //        }
 //    }
 
-    @PostMapping("/reviews/{id}/edit")
-        public String submitEdit(@ModelAttribute BookReview editedReview){
-            BookReview postToEdit = bookReviewDao.getById(editedReview.getId());
-            postToEdit.setTitle(editedReview.getTitle());
-            postToEdit.setBody(editedReview.getBody());
-            bookReviewDao.save(editedReview);
-            return "redirect:/reviews";
-        }
 
+//    @GetMapping("reviews/{id}/user/{reviewid}/edit")
+//    public String editReview(@PathVariable long id, @PathVariable long reviewid, Model model) {
+//        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+//        BookReview reviewToEdit = bookReviewDao.getById(reviewid);
+//        model.addAttribute("editReview",bookReviewDao.getById(id));
+//        return "/reviews/edit";
+//    }
+
+
+    //    mike
      @PostMapping("/reviews/{id}/delete")
      public String deleteReview(@PathVariable long id) {
-        bookReviewDao.deleteById(id);
+         User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+         BookReview reviewToDelete = bookReviewDao.getById(id);
+         if(reviewToDelete.getUser().getId() == currentUser.getId()){
+             bookReviewDao.deleteById(id);
+         }
+
         return "redirect:/reviews";
      }
 
 
 
+
+// this one
+//    @GetMapping("/reviews/{id}")
+//    public String showUserReview(@PathVariable long id, Model model){
+//        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+//        BookReview showReview = bookReviewDao.getById(id);
+//        boolean isOwner = false;
+//        if(showReview.getUser().getId() == currentUser.getId()){
+//            isOwner = true;
+//        }
+//        model.addAttribute("isOwner", isOwner);
+//        model.addAttribute("review", bookReviewDao.getById(id));
+//        model.addAttribute("showBook", showReview);
+//        return "/reviews/individual-review";
+//    }
